just some more things
