+++ RaspberryPie/flask_backend.py
@@ -1,0 +1,6 @@
+from flask import Flask, render_template, make_response
+
+app = Flask(__name__)
+@app.route("/")
+def index():
+    pass
